수정보완 (#54)
* test: Like 도메인 테스트코드 수정보완

* test: User 도메인 테스트코드 수정보완

* refactor: S3Service 리팩토링

* feat: Mock S3Service 설정

* fix: follower, following count 메소드 추가

* refactor: test/config 폴더명 변경

* fix: user, category 삭제시 연관된 엔티티 soft delete

+++ src/main/java/org/ahpuh/surf/user/dto/UserDto.java
@@ -20,9 +20,9 @@ public class UserDto {
 
     private String url;
 
-    private int followerCount;
+    private long followerCount;
 
-    private int followingCount;
+    private long followingCount;
 
     private Boolean accountPublic;
